requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==2.3.7
elasticsearch==8.10.0
requests==2.31.0
geopy==2.4.0